pyproject.toml
[tool]
[tool.poetry]
name = "deluge-grpc"
version = "0.1.0"
homepage = "https://github.com/chrisbc/deluge-grpc"
description = "gRPC interface to deluge-card."
authors = ["Chris B Chamberlain <chrisbc@artisan.co.nz>"]
readme = "README.md"
license =  "GPL-3.0-only"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "deluge_grpc" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dev-dependencies]
black  = { version = ">=22.3"}
bump2version = {version = "^1.0.1", optional = true}
click = "^8.1.3"
flake8  = { version = "^3.9.2"}
flake8-docstrings = { version = "^1.6.0"}
isort  = { version = "^5.8.0"}
jinja2 = {version = "==3.0.3"}
mkdocs  = {version = ">=1.1"}
mkdocs-autorefs = {version = ">=0.3.1"}
mkdocs-include-markdown-plugin  = {version = ">=1.0.0"}
mkdocs-material  = {version = ">=6.1.7"}
mkdocs-material-extensions  = {version = ">=1.0.1"}
mkdocs-pymdownx-material-extras = "^2.0.3"
mkdocstrings  = {version = ">=0.18.0"}
mkdocstrings-python = ">=0.6.5"
mypy = {version = "^0.900"}
pip  = { version = "^20.3.1", optional = true}
pre-commit = {version = "^2.12.0", optional = true}

#pymdown-extensions = ">=9.4"
pytest  = { version = "^6.2.4"}
pytest-cov  = { version = "^2.12.0"}
toml = {version = "^0.10.2", optional = true}
tox  = { version = "^3.20.1"}
twine  = {version = "^3.3.0"}
virtualenv  = { version = "^20.2.2"}
types-protobuf = "^3.19.21"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
deluge-card = "^0.6.1"
grpcio = "^1.46.3"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "flake8-docstrings",
    "pytest-cov"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs"
    ]

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py38','py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
